Refactor: Restructure CLI deployment utilities and deduplicate options across Cloud Run and Agent Engine deployment modules
