+++ dev/src/cli/deploy/cli_deploy_cloud_run.ts
@@ -6,38 +6,24 @@
 import * as fs from 'node:fs/promises';
 import * as path from 'node:path';
 
-import {AgentFileOptions, AgentLoader} from '../../utils/agent_loader.js';
-import {isFile, isFolderExists, saveToFile} from '../../utils/file_utils.js';
+import {AgentLoader} from '../../utils/agent_loader.js';
+import {isFile, isFolderExists} from '../../utils/file_utils.js';
 import {
+  BaseDeployOptions,
+  CreateDockerFileContentOptions,
   copyAgentFiles,
+  createDockerFile,
+  createDockerFileContent,
   createPackageJson,
   resolveDefaultFromGcloudConfig,
   spawnAsync,
 } from './deploy_utils.js';
 
-export interface CreateDockerFileContentOptions {
-  appName?: string;
-  project: string;
-  region?: string;
-  port: number;
-  withUi: boolean;
-  logLevel: string;
-  allowOrigins?: string;
-  sessionServiceUri?: string;
-  artifactServiceUri?: string;
-  otelToCloud?: boolean;
-  a2a?: boolean;
-}
+export {createDockerFileContent, type CreateDockerFileContentOptions};
 
-export interface DeployToCloudRunOptions extends CreateDockerFileContentOptions {
-  agentPath: string;
+export interface DeployToCloudRunOptions extends BaseDeployOptions {
   serviceName: string;
-  tempFolder: string;
-  adkVersion: string;
   extraGcloudArgs?: string[];
-  otelToCloud?: boolean;
-  agentFileLoadOptions?: AgentFileOptions;
-  a2a?: boolean;
 }
 
 function validateGcloudExtraArgs(
@@ -109,86 +95,6 @@ function prepareGCloudArguments(options: DeployToCloudRunOptions): string[] {
   gcloudCommands.push(...extraArgsWithoutLabels);
 
   return gcloudCommands;
-}
-
-export function createDockerFileContent(
-  options: CreateDockerFileContentOptions,
-): string {
-  const adkCommand = options.withUi ? 'web' : 'api_server';
-  const adkServerOptions = [`--port=${options.port}`, '--host=0.0.0.0'];
-
-  if (options.logLevel) {
-    adkServerOptions.push(`--log_level=${options.logLevel}`);
-  }
-
-  if (options.allowOrigins) {
-    adkServerOptions.push(`--allow_origins=${options.allowOrigins}`);
-  }
-
-  if (options.artifactServiceUri) {
-    adkServerOptions.push(
-      `--artifact_service_uri=${options.artifactServiceUri}`,
-    );
-  }
-
-  if (options.sessionServiceUri) {
-    adkServerOptions.push(`--session_service_uri=${options.sessionServiceUri}`);
-  }
-
-  if (options.otelToCloud) {
-    adkServerOptions.push('--otel_to_cloud');
-  }
-
-  if (options.a2a) {
-    adkServerOptions.push('--a2a');
-  }
-
-  return `
-FROM node:lts-alpine
-WORKDIR /app
-
-# Create a non-root user
-RUN adduser --disabled-password --gecos "" myuser
-
-# Switch to the non-root user
-USER myuser
-
-# Set up environment variables - Start
-ENV PATH="/home/myuser/.local/bin:$PATH"
-ENV GOOGLE_GENAI_USE_VERTEXAI=1
-ENV GOOGLE_CLOUD_PROJECT=${options.project}
-ENV GOOGLE_CLOUD_LOCATION=${options.region}
-# Set up environment variables - End
-
-# Copy application files
-COPY --chown=myuser:myuser "agents/${options.appName}/" "/app/agents/${
-    options.appName
-  }/"
-COPY --chown=myuser:myuser "package.json" "/app/package.json"
-COPY --chown=myuser:myuser "package-lock.json" "/app/package-lock.json"
-COPY --chown=myuser:myuser "node_modules" "/app/node_modules"
-# Copy application files
-
-# Install Agent Deps - Start
-RUN npm install @google/adk-devtools@latest
-RUN npm install --production
-# Install Agent Deps - End
-
-EXPOSE ${options.port}
-
-CMD npx adk ${adkCommand} /app/agents/${options.appName} ${adkServerOptions.join(
-    ' ',
-  )}`;
-}
-
-async function createDockerFile(
-  targetFolder: string,
-  options: CreateDockerFileContentOptions,
-) {
-  const dockerFilePath = path.join(targetFolder, 'Dockerfile');
-  await saveToFile(dockerFilePath, createDockerFileContent(options));
-
-  console.info('Creating Dockerfile complete:', dockerFilePath);
 }
 
 export async function deployToCloudRun(options: DeployToCloudRunOptions) {
