+++ dev/src/cli/deploy/deploy_utils.ts
@@ -7,7 +7,7 @@ import {exec, spawn, SpawnOptions} from 'node:child_process';
 import * as fs from 'node:fs/promises';
 import * as path from 'node:path';
 import {promisify} from 'node:util';
-import {AgentLoader} from '../../utils/agent_loader.js';
+import {AgentFileOptions, AgentLoader} from '../../utils/agent_loader.js';
 import {
   loadFileData,
   saveToFile,
@@ -34,6 +34,107 @@ export const spawnAsync = (
 };
 
 export const REQUIRED_NPM_PACKAGES = ['@google/adk'];
+
+export interface CreateDockerFileContentOptions {
+  appName?: string;
+  project: string;
+  region?: string;
+  port: number;
+  withUi: boolean;
+  logLevel: string;
+  allowOrigins?: string;
+  sessionServiceUri?: string;
+  artifactServiceUri?: string;
+  otelToCloud?: boolean;
+  a2a?: boolean;
+}
+
+export interface BaseDeployOptions extends CreateDockerFileContentOptions {
+  agentPath: string;
+  tempFolder: string;
+  adkVersion: string;
+  agentFileLoadOptions?: AgentFileOptions;
+}
+
+export function createDockerFileContent(
+  options: CreateDockerFileContentOptions,
+): string {
+  const adkCommand = options.withUi ? 'web' : 'api_server';
+  const adkServerOptions = [`--port=${options.port}`, '--host=0.0.0.0'];
+
+  if (options.logLevel) {
+    adkServerOptions.push(`--log_level=${options.logLevel}`);
+  }
+
+  if (options.allowOrigins) {
+    adkServerOptions.push(`--allow_origins=${options.allowOrigins}`);
+  }
+
+  if (options.artifactServiceUri) {
+    adkServerOptions.push(
+      `--artifact_service_uri=${options.artifactServiceUri}`,
+    );
+  }
+
+  if (options.sessionServiceUri) {
+    adkServerOptions.push(`--session_service_uri=${options.sessionServiceUri}`);
+  }
+
+  if (options.otelToCloud) {
+    adkServerOptions.push('--otel_to_cloud');
+  }
+
+  if (options.a2a) {
+    adkServerOptions.push('--a2a');
+  }
+
+  return `
+FROM node:lts-alpine
+WORKDIR /app
+
+# Create a non-root user
+RUN adduser --disabled-password --gecos "" myuser
+
+# Switch to the non-root user
+USER myuser
+
+# Set up environment variables - Start
+ENV PATH="/home/myuser/.local/bin:$PATH"
+ENV GOOGLE_GENAI_USE_VERTEXAI=1
+ENV GOOGLE_CLOUD_PROJECT=${options.project}
+ENV GOOGLE_CLOUD_LOCATION=${options.region}
+# Set up environment variables - End
+
+# Copy application files
+COPY --chown=myuser:myuser "agents/${options.appName}/" "/app/agents/${
+    options.appName
+  }/"
+COPY --chown=myuser:myuser "package.json" "/app/package.json"
+COPY --chown=myuser:myuser "package-lock.json" "/app/package-lock.json"
+COPY --chown=myuser:myuser "node_modules" "/app/node_modules"
+# Copy application files
+
+# Install Agent Deps - Start
+RUN npm install @google/adk-devtools@latest
+RUN npm install --production
+# Install Agent Deps - End
+
+EXPOSE ${options.port}
+
+CMD npx adk ${adkCommand} /app/agents/${options.appName} ${adkServerOptions.join(
+    ' ',
+  )}`;
+}
+
+export async function createDockerFile(
+  targetFolder: string,
+  options: CreateDockerFileContentOptions,
+) {
+  const dockerFilePath = path.join(targetFolder, 'Dockerfile');
+  await saveToFile(dockerFilePath, createDockerFileContent(options));
+
+  console.info('Creating Dockerfile complete:', dockerFilePath);
+}
 
 export async function copyAgentFiles(
   agentLoader: AgentLoader,
